+++ NEWS.md
@@ -2,6 +2,7 @@ PENDING: Dataiter 0.50
 ======================
 
 * `ListOfDicts.drop_na`: New method
+* `ListOfDicts.keys`: New method
 
 2023-11-08: Dataiter 0.49
 =========================